Initial separation of frameskipping and step single frame logic into frame_manager
