+++ src/emu/emufwd.h
@@ -246,4 +246,7 @@ class validity_checker;
 // declared in video.h
 class video_manager;
 
+// declared in frame.h
+class frame_manager;
+
 #endif // MAME_EMU_EMUFWD_H
