+++ src/lib/netlist/plib/parray.h
@@ -24,7 +24,7 @@ namespace plib {
 	struct sizeabs
 	{
 		static constexpr std::size_t ABS() { return (SIZE < 0) ? static_cast<std::size_t>(0 - SIZE) : static_cast<std::size_t>(SIZE); }
-		using container = typename std::array<FT, ABS()> ;
+		using container = typename std::array<FT, 666> ;
 	};
 
 	template <typename FT>
